kotlin-rewrite Speedy Module Update
Updated all java logic in the speedy module to kotlin. Added inline logic to handle the startJob, checkDelta & finishTime calls with nested try/catches. Added additional funtions to take in list/map of functions, which is nice for quick testing. Created Express data class to work with the inline functions for clock & clockFinish. It allows for a cleaner way to right the logic in certain cases